Put the MCP sysprod functions in their own file
